testing: move MovieRental exercises to the additional_exercises file
The MovieRental part 1 exercise will be replaced with a new exercise
which is faster.
The students did not manage to end the MovieRental part 1 exercise
within 2 hours.
Thus an exercise is needed which shows TDD with multiple classes
in less time.
MovieRental part 2 was always for people who were faster and is moved
to the additional exercises without replacement.